Generate digital links after order is completed, not after saving line items

+++ app/models/spree/line_item_decorator.rb
@@ -1,25 +1,24 @@
 Spree::LineItem.class_eval do
   
   has_many :digital_links, :dependent => :destroy
-  after_save :create_digital_links, :if => :digital?
+  # after_save :create_digital_links, :if => :digital?
   
   def digital?
     variant.digital?
   end
   
-  private
-  
-  # TODO there is no reason to create the digital links until the order is complete
   # TODO: PMG - Shouldn't we only do this if the quantity changed?
   def create_digital_links
-    digital_links.delete_all
+    if self.digital? && self.order.complete?
+      digital_links.delete_all
 
-    #include master variant digitals
-    master = variant.product.master
-    if(master.digital?)
-      create_digital_links_for_variant(master)
+      #include master variant digitals
+      master = variant.product.master
+      if(master.digital?)
+        create_digital_links_for_variant(master)
+      end
+      create_digital_links_for_variant(variant) unless variant.is_master
     end
-    create_digital_links_for_variant(variant) unless variant.is_master
   end
 
   def create_digital_links_for_variant(variant)
